+++ resources/views/components/ui/dropdown-menu-content.blade.php
@@ -14,8 +14,9 @@
     $anchorAttr = 'x-anchor.'.$placement.'.offset.'.$sideOffset.'="'.$anchorRef.'"';
 @endphp
 
-<template x-teleport="body">
+<template x-teleport="body" wire:ignore>
     <div
+        x-blat-dialog-layer
         x-show="open"
         x-cloak
         x-ref="menu"